+++ Source/UI/EditAboutView.swift
@@ -9,11 +9,11 @@
 import Foundation
 import UIKit
 
-class EditAboutView: UIView, Saveable, UITextViewDelegate {
+class EditAboutView: UIView, Saveable, UITextFieldDelegate {
 
     lazy var nameView: EditValueView = {
         let view = EditValueView(label: .name)
-        view.textView.delegate = self
+        view.textField.addTarget(self, action: #selector(textFieldDidChange), for: .editingChanged)
         view.backgroundColor = .cardBackground
         view.textView.backgroundColor = .cardBackground
         return view
@@ -21,7 +21,6 @@ class EditAboutView: UIView, Saveable, UITextViewDelegate {
 
     lazy var bioView: EditValueView = {
         let view = EditValueView(label: .bio)
-        view.textView.delegate = self
         view.backgroundColor = .cardBackground
         view.textView.backgroundColor = .cardBackground
         return view
@@ -29,22 +28,45 @@ class EditAboutView: UIView, Saveable, UITextViewDelegate {
 
     init() {
         super.init(frame: .zero)
+        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillShow), name: UIResponder.keyboardWillShowNotification, object: nil)
+        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillHide), name: UIResponder.keyboardWillHideNotification, object: nil)
+        
         self.backgroundColor = .cardBackground
         let spacer = Layout.addSpacerView(toTopOf: self)
         spacer.backgroundColor = .appBackground
-        var separator = Layout.addSeparator(southOf: spacer)
+        let separator = Layout.addSeparator(southOf: spacer)
         Layout.fillSouth(of: separator, with: self.nameView)
-        separator = Layout.addSeparator(southOf: self.nameView)
-        Layout.fillSouth(of: separator, with: self.bioView)
+        Layout.fillSouth(of: self.nameView, with: self.bioView)
+        bioView.bottomAnchor.constraint(equalTo: self.bottomAnchor).isActive = true
     }
+    
+    deinit {
+         NotificationCenter.default.removeObserver(self)
+       }
 
     required init?(coder aDecoder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
 
     func update(with about: About) {
-        self.nameView.textView.text = about.name
-        self.bioView.textView.text = about.description
+        self.nameView.text = about.name ?? ""
+        self.bioView.text = about.description ?? ""
+    }
+    
+    @objc func keyboardWillShow(notification: Notification) {
+        guard let keyboardFrame = notification.keyboardFrameEnd else {
+            return
+        }
+        let bottomOfTextView = bioView.convert(bioView.bounds, to: self).maxY
+        let visibleArea = self.frame.height - keyboardFrame.height
+        
+        if bottomOfTextView > visibleArea {
+            bioView.bottomConstraint?.constant = -keyboardFrame.height - Layout.verticalSpacing
+        }
+    }
+    
+    @objc func keyboardWillHide(notification: Notification) {
+        bioView.bottomConstraint?.constant = -Layout.verticalSpacing
     }
 
     // MARK: First responding
@@ -55,20 +77,20 @@ class EditAboutView: UIView, Saveable, UITextViewDelegate {
     }
 
     func resignFirstResponders() {
-        self.nameView.textView.resignFirstResponder()
+        self.nameView.textField.resignFirstResponder()
         self.bioView.textView.resignFirstResponder()
     }
 
-    // MARK: UITextViewDelegate
+    // MARK: UITextField editing changed action
 
-    func textViewDidChange(_ textView: UITextView) {
+    @objc private func textFieldDidChange(_ textField: UITextField) {
         self.delegate?.saveable(self, isReadyToSave: self.isReadyToSave)
     }
 
     // MARK: Saveable
 
     var isReadyToSave: Bool {
-        self.nameView.textView.text.isValidName
+        self.nameView.text.isValidName
     }
 
     // TODO bah protocol requires this, wish is could be defaulted somehow